+++ __tests__/Car.test.js
@@ -1,0 +1,5 @@
+describe("테스트", () => {
+  test("테스트1", () => {
+    expect(true);
+  });
+});
